Clean merge item pair.

Called in get_metadata to add additional rows with merge item pairs. This is needed so that all these 'other' columns are preserved when merging data with metadata.

Usage

clean_merge_pair_metadata(
  data,
  merge_col = "merge_with",
  suffix_to_add = "_ITEM_TO_MERGE_WITH_PAIR"
)

Arguments

data

A data frame with metadata

merge_col

A character vector with the column name in which the variable to merge with is defined.

suffix_to_add

A Character vector with the suffix to add to the item_name. This is used to detect item pairs and merge them when in merge_meta_with_data().