Clinical Trial test data

A data.frame containing randomly created clinical trial data. Used for testing purposes. It will also be used to run the app with example data when executing run_app() with the default configuration settings, or when running either test_clinsight() or golem::run_dev(). See also vignette('Metadata') and vignette('clinsight') for an explanation on how to create custom study data for clinsight.

Usage

clinsightful_data

Format

a data.frame with 6076 rows and 22 variables.

site_code: character or integer, identifier for study site.

subject_id: character, unique identifier for a subject.

event_id: character, event identifier in raw data. Replaced by event_name and event_label within ClinSight. An "event" generally characterizes some sort of site visit (e.g. "Screening", "Visit 1", "Visit 2"). Note that some event_ids track events that could apply outside of any visit, like AE, ConMed, Medical History, etc.

event_name: character, the name of an event identifier used within the application. Created after merging raw data with metadata.

event_label: character, shorter name of an event identifier. Created after merging raw data with metadata. Used instead of event_name for some compact visualizations.

event_date: Date, the date associated with event_name.

event_repeat: integer, keeps track of unique event_id for a single subject_id and event_date.

form_id: character, a unique form_id identifier. Currently unused; item_group is used instead, which is controlled using the metadata.

form_repeat: integer, keeps track of unique item_names collected from a specific form for a given subject_id. form_repeat is particularly helpful when consolidating data like Adverse Events into this data format. Specifically, if more than one AE is collected on a patient, they'll have more than one form_repeat.

edit_date_time: date-time (POSIXct), the last time this record was edited.

region: character, describing the region code that site_code falls under.

day: a difftime() R object, measuring the number of days each visit is from screening.

vis_day: numeric, a numeric representation of day. Created after merging raw data with metadata.

vis_num: numeric, a numeric representation of event_name. Created after merging raw data with metadata. Only counted for scheduled events (can be controlled in the metadata).

item_name: character, the name of a metric or parameter of interest.

item_type: character, classifies item_names into either 'continuous' or 'other', where continuous types are those generally associated with the CDISC "basic data structure" (BDS). That is, each item_name metric is collected over time at a patient visit (event_name). The 'other' type represents all non-time dependent measures, like demographic info, adverse events, Medications, medical history, etc.

item_group: character, defines in which form the items with item_name will appear.

item_value: character, the measurement collected for a given item_name. The value collected may be a number like 150 (when collecting a patient's weight) or a word (such as 'white' for the subject's race).

item_unit: character, tracking the unit of measurement for item_name and item_value.

lower_lim: numeric, some item_names (particularly the 'continuous' type) have a pre-defined range of values that are considered normal. This is the lower limit to that range.

upper_lim: numeric, some item_names (particularly the 'continuous' type) have a pre-defined range of values that are considered normal. This is the upper limit to that range.

significance: character, either 'CS' which means 'Clinically Significant' or 'NCS' which means 'Not Clinically Significant'.

reason_notdone: character, an effort to describe why the item_value field is NA / missing.

Source

Can be recreated with get_metadata(get_raw_csv_data(data_path = system.file("raw_data", package = "clinsight"))).