Collect query data

Collects query data from a remote database.

Usage

collect_query_data(
  db_path,
  db_table = "query_data",
  time_var = "timestamp",
  query_id = "query_id",
  query_follow_up_number = "n",
  resolved_var = "resolved",
  resolved_date_var = "resolved_date"
)

Arguments

db_path

A character string with the file path to the database.

db_table

Character string. Name of the table to collect. Will only be used if data is a character string to a database.

time_var

Character string with the name of the timestamp variable.

query_id

Character string with the name of the query id variable

query_follow_up_number

Character string. Follow-up number of the query. Will be incremented for follow-up message on the initial query.

resolved_var

Character string with the name of the variable that shows whether the query is resolved ('Yes') or not ('No').

resolved_date_var

Character string. Date-time when the query was marked as resolved.

Value

A data frame collected from a remote database.