Create Report

Creates a PDF markdown report.

Usage

create_report(fileinput, reviewer, study_sites, review_df, query_df)

Arguments

fileinput

Character vector with the name of the file to be created.

reviewer

Character vector with the name of the current reviewer.

study_sites

Character vector with the names of the sites that were reviewed.

review_df

Data frame with review data to be used in the report.

query_df

Data frame with query data to be used in the report.

Value

A PDF report will be created.