Create app database

Creates application database. To create a database with all data flagged as 'new', use the default settings of reviewed, reviewer, and status.

Usage

db_create(data, db_path, reviewed = "No", reviewer = "", status = "new")

Arguments

data

A data frame with review data (Usually created with get_review_data()).

db_path

A character vector with the path to the database to be created.

reviewed

Character vector. Sets the reviewed tag in the review database.

reviewer

Character vector. Sets the reviewer in the review database.

status

Character vector. Sets the status in the review database. Defaults to new.

Value

A database will be created. Nothing else will be returned.

See also