Get raw data from CSV files

Get raw data from CSV files

Usage

get_raw_csv_data(
  data_path = Sys.getenv("RAW_DATA_PATH"),
  synch_time = time_stamp(),
  exclude = c("README.csv$", "Pending_forms.csv$", "MEDRA.csv$", "WHODrug.csv$",
    "_Queries.csv$"),
  delim = ",",
  skip = 1
)

Arguments

data_path

Path to the folder that contains the CSV files with the application data.

synch_time

Time at which the data was extracted from the EDC system. Defaults to the current date time. Important to set this correctly, since it will be shown in the application. By default, a warning will be given in the application if the synchronization time is more than one day old.

exclude

character vector with regular expressions that identify csv files that should be excluded from the study data. Useful to exclude files with different data structures, or files with metadata.

delim

Delimiter to use to read in files.

skip

Number of rows to skip when reading in files.

Value

A data frame with raw application data.