Get package test results

Creates package results and wraps it in a list, together with a timestamp at which the results were created and the current sessionInfo.

Usage

get_test_results(
  package = "clinsight",
  reporter = testthat::ListReporter,
  outfile = NULL,
  run_as_not_cran = TRUE
)

Arguments

package

Character vector with the poackage name.

reporter

a testthat file reporter. defaults to testthat::ListReporter.

outfile

If provided, the results will be saved in a .rds file on the

run_as_not_cran

Logical. If TRUE, will provide a temporary environmental variable NOT_CRAN='true' to ensure that tests are not skipped.

Value

A list with the outcome. If outfile is provided, the result will also be saved in the location of outfile.