Navigate forms - Shiny module UI

Shiny UI module containing the UI needed to navigate between forms. The UI contains a back and forward which changes the active form. In addition, the active form is displayed either in bold or in normal font face, depending on whether the form contains any non-reviewed data.

Usage

mod_navigate_forms_ui(id)

Arguments

id

Character string, used to connect the module UI with the module Server.

See also