Queries - Shiny module UI

A Shiny module. Used to display all raised queries.

Usage

mod_queries_ui(id)

Arguments

id

Character string, used to connect the module UI with the module Server.

Details

The module displays all raised queries in a table. Upon clicking on a query, the query follow-up messages will be shown. In addition, it will be possible to write a follow-up message. The logic for writing the follow-up message is extracted in a different module, named mod_query_follow_up_server(). New queries are created in the module mod_query_add_server(). The button to initiate this module can be found in the main sidebar mod_main_sidebar_server().

See also