Review form - Shiny module Server

A shiny module. Used to mark data as reviewed and add comments to data of a specific form/page in the database and in the internal application data.

This module is used to display whether the form that is actively being displayed in the application has any values that are newly entered or updated since the last review session. In addition, if a page contains new data, the user can mark all data in the form as being reviewed and can (optionally) add a comment to this review action. The data will be saved in a database. All review activity is stored with an audit-trail, with date/time stamps and with the reviewer's name.

Usage

mod_review_forms_server(
  id,
  r,
  active_form,
  active_tab,
  review_required_data,
  db_path
)

Arguments

id

Character string, used to connect the module UI with the module Server.

r

Common reactive values. Used to access subject_id, user_name, review_data, and status_data. The module can change data in the review_data and to status_data data frames.

active_form

Reactive value. Used to determine which form is active and whether that form needs review.

active_tab

Reactive value containing the active tab. Needed to hide the review controls if non-relevant tabs are selected (tabs that do not contain common forms or study forms).

review_required_data

a data frame with information about whether review is mandatory for each form. Should contain the columns item_group and review_required.

db_path

Character string with the file path to the database.

See also