Start page - Shiny module Server

A Shiny module. Start page of the main application.

Usage

mod_start_page_server(id, r, rev_data, navinfo, all_forms, table_names)

Arguments

id

Character string, used to connect the module UI with the module Server.

r

Common reactive values. Used to send back information about the selected subject id (r$subject_id) to the main server. In addition, r$filtered_subjects is needed to verify whether an ID selected in the tablerow is in the active, filtered selection.

rev_data

Common reactiveValues(). Used for navigation purposes, to set the subject for which we want to browse the review data from.

navinfo

Common reactive values. Used to send back the request for the change in form/subject id back to the server. Contains the character strings active_form, active_tab, and trigger_page_change. The module will create these variables if they are not yet available in navinfo.

all_forms

A data frame containing two columns: a column named "form" with all the form to navigate through in the correct order, and a column named "main_tab", showing the name of the main_tab in which the form can be found.

table_names

An optional character vector. If provided, will be used within datatable_custom(), to improve the column names in the final interactive tables.

Details

The start page contains navigation buttons and a table that shows one row per patient. The table shows general study information such as study status (active or inactive participant), main diagnosis, age, sex, number of conducted study visits, and the patient's data status (new data, updated data, or old data). In addition, in the module there are two navigation options. First, the user can navigate to the selected patient directly. Secondly, the user can indirectly call the module mod_navigate_review_server() to open a modal in which all the forms are shown that need review of the selected patient. navigate to patient that is currently highlighted. highlighted to a participant study and options to navigate directly to the selected forms/patients n the overview tables.

See also