Application Metadata

A list of data frames and settings containing metadata that will be used for the application. The metadata will be merged with with raw data. It controls the variables that will be used in the application, and in which tab the variables will be shown. The goal is that most, if not all, study-specific data will be captured in the metadata, leaving the scripts to run the application largely unaltered between studies.See vignette("Metadata") for in-depth information.

Usage

metadata

Format

metadata: A list with 9 objects.

str(metadata)
#> List of 9
#>  $ column_names   : tibble [10 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ name_raw: chr [1:10] "SiteCode" "SubjectId" "EventId" "EventDate" ...
#>   ..$ name_new: chr [1:10] "site_code" "subject_id" "event_id" "event_date" ...
#>  $ events         : tibble [10 x 10] (S3: tbl_df/tbl/data.frame)
#>   ..$ event_id               : chr [1:10] "SCR" "VIS1" "VIS2" "VIS3" ...
#>   ..$ event_id_pattern       : chr [1:10] "^SCR$" "^VIS1$" "^VIS2$" "^VIS3$" ...
#>   ..$ is_regular_visit       : logi [1:10] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ event_label_custom     : chr [1:10] NA "V1" "V2" "V3" ...
#>   ..$ event_name_custom      : chr [1:10] "Screening" "Visit 1" "Visit 2" "Visit 3" ...
#>   ..$ is_baseline_event      : logi [1:10] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ generate_labels        : logi [1:10] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ meta_event_order       : int [1:10] 1 2 3 4 5 6 7 8 9 10
#>   ..$ add_visit_number       : logi [1:10] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ add_event_repeat_number: logi [1:10] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ common_forms   :'data.frame':  44 obs. of  4 variables:
#>   ..$ item_name : chr [1:44] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ item_type : chr [1:44] "other" "other" "other" "other" ...
#>   ..$ item_group: chr [1:44] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>   ..$ merge_with: chr [1:44] NA NA NA NA ...
#>  $ study_forms    :'data.frame':  29 obs. of  6 variables:
#>   ..$ item_name  : chr [1:29] "Systolic blood pressure" "Diastolic blood pressure" "Pulse" "Resp" ...
#>   ..$ item_type  : chr [1:29] "continuous" "continuous" "continuous" "continuous" ...
#>   ..$ item_group : chr [1:29] "Vital signs" "Vital signs" "Vital signs" "Vital signs" ...
#>   ..$ unit       : chr [1:29] "mmHg" "mmHg" "beats/min" "breaths/min" ...
#>   ..$ lower_limit: chr [1:29] "90" "55" "60" "12" ...
#>   ..$ upper_limit: chr [1:29] "160" "90" "100" "20" ...
#>  $ general        :'data.frame':  20 obs. of  3 variables:
#>   ..$ item_name : chr [1:20] "Age" "Sex" "ECOG" "Eligible" ...
#>   ..$ item_type : chr [1:20] "other" "other" "other" "other" ...
#>   ..$ item_group: chr [1:20] "General" "General" "General" "General" ...
#>  $ form_level_data:'data.frame':  11 obs. of  4 variables:
#>   ..$ item_group         : chr [1:11] "Adverse events" "Medication" "Conc. Procedures" "Medical History" ...
#>   ..$ item_scale         : logi [1:11] NA NA NA NA FALSE TRUE ...
#>   ..$ use_unscaled_limits: logi [1:11] NA NA NA NA TRUE FALSE ...
#>   ..$ review_required    : logi [1:11] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ table_names    : tibble [15 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ table_name: chr [1:15] "Edit date" "Date" "Event" "Event" ...
#>   ..$ raw_name  : chr [1:15] "edit_date_time" "event_date" "event_label" "event_name" ...
#>  $ settings       :List of 4
#>   ..$ pre_pivot_fns  : chr "apply_study_specific_suffix_fixes"
#>   ..$ post_pivot_fns : chr "apply_edc_specific_changes"
#>   ..$ post_merge_fns : chr "apply_study_specific_fixes"
#>   ..$ treatment_label: chr "💊 Tₓ"
#>  $ items_expanded : tibble [220 x 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ form_type  : chr [1:220] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ var        : chr [1:220] "AE_AESPID" "AE_AETERM" "AE_AESI" "AE_AESTDAT" ...
#>   ..$ suffix     : chr [1:220] NA NA NA NA ...
#>   ..$ item_name  : chr [1:220] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ item_type  : chr [1:220] "other" "other" "other" "other" ...
#>   ..$ item_group : chr [1:220] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>   ..$ unit       : chr [1:220] NA NA NA NA ...
#>   ..$ lower_limit: chr [1:220] NA NA NA NA ...
#>   ..$ upper_limit: chr [1:220] NA NA NA NA ...

Source

Can be created with an Excel file. The Excel file format is chosen so that the metadata can be changed easily per study. See raw-data/metadata.R for details.